cleaning the parser code and adding a documentation extractor
