+++ arkdoc/parser/__init__.py
@@ -1,4 +1,3 @@
 #!/usr/bin/env python3
 
-from .ast import NodeType, Node
 from .parser import Parser
