+++ arkdoc/parser/parser.py
@@ -1,88 +1,45 @@
 #!/usr/bin/env python3
 
 import re
-from typing import List, NamedTuple
+from typing import List
+from dataclasses import dataclass
+from collections.abc import Iterable
 
 from .. import logger
+from .tokenizer import tree_from_tokens, tokenize, Token
 
 
-class Token(NamedTuple):
-    type: str
-    value: str
-    line: int
-    column: int
+def deep_flatten(lst):
+    return ([a for i in lst for a in
+             deep_flatten(i)] if isinstance(lst, Iterable) else [lst])
+
+
+@dataclass
+class Documentation:
+    comments: List[Token]
+    target: List
+
+    def _token_format(self, token: Token):
+        return token.value
+
+    def extract_signature(self, on: List[Token] = None):
+        top = self.target[:] if on is None else on
+        while True:
+            if isinstance(top, list) and \
+                    len(top) and isinstance(top[0], list):
+                top = top[0]
+            else:
+                break
+
+        transform = lambda L: list(map(lambda t: t.value, deep_flatten(L)))
+
+        if on is None:
+            return transform(top[:2]) + self.extract_signature(on=top[2])
+        else:
+            return transform(top[1:2])
 
     def __str__(self):
-        return f"Token({self.type}, '{self.value}', line={self.line}, col={self.column})"
-
-
-Keywords = "let mut set del fun if while import begin quote".split()
-TokenSpecification = [
-    ('NUMBER', r'\d+(\.\d*)?'),
-    ('STRING', r'"[^"]*"'),
-    ('ID', r'[\w:?=!@&<>+\-%*/.]+'),
-    ('LPAREN', r'[(\[{]'),
-    ('RPAREN', r'[)\]}]'),
-    ('COMMENT', r'#[^\n]*'),
-    ('NEWLINE', r'\n'),
-    ('SKIP', r'[ \t]+'),
-    ('MISMATCH', r'.'),
-]
-
-
-def tokenize(code: str) -> List[Token]:
-    tok_regex = '|'.join(
-        '(?P<%s>%s)' %
-        pair for pair in TokenSpecification
-    )
-    line_num = 1
-    line_start = 0
-
-    lines = code.split('\n')
-
-    for mo in re.finditer(tok_regex, code):
-        kind = mo.lastgroup
-        value = mo.group()
-        column = mo.start() - line_start
-
-        if kind == 'ID' and value in Keywords:
-            kind = value
-        elif kind == 'NEWLINE':
-            line_start = mo.end()
-            line_num += 1
-            continue
-        elif kind == 'SKIP':
-            continue
-        elif kind == 'MISMATCH':
-            raise RuntimeError(
-                f'{value!r} unexpected on line {line_num}\n{lines[line_num - 1]}'
-            )
-        yield Token(kind, value, line_num, column)
-
-
-
-def tree_from_tokens(tokens: List[Token]) -> List:
-    if len(tokens) == 0:
-        raise SyntaxError('unexpected EOF')
-
-    token = tokens.pop(0)
-
-    L = []
-    while token.type == 'COMMENT':
-        L.append(token)
-        token = tokens.pop(0)
-
-    if token.type == 'LPAREN':
-        L2 = []
-        while tokens[0].type != 'RPAREN':
-            L2.append(tree_from_tokens(tokens))
-        tokens.pop(0)
-        L.append(L2)
-        return L
-    elif token.type == 'RPAREN':
-        raise SyntaxError(f"unexpected ) on line {token.line}, at {token.column}")
-    else:
-        return token
+        return f"Documentation({len(self.comments)} comments, {self.extract_signature()})"
 
 
 class Parser:
@@ -97,9 +54,35 @@ class Parser:
             else:
                 self.visit(n, depth + 1)
 
+    def _is_doc_comment(self, token: Token) -> bool:
+        return token.type == 'COMMENT' and re.match(
+            r'^#+ *@\w+', token.value
+        ) is not None
+
+    def extract_doc(self, node: List):
+        doc = None
+
+        for n in node:
+            if not isinstance(n, list):
+                if self._is_doc_comment(n):
+                    if doc is None:
+                        doc = Documentation([], None)
+                    doc.comments.append(n)
+            elif doc is None:
+                self.extract_doc(n)
+            else:
+                doc.target = n
+                yield doc
+                doc = None
+
+                yield from self.extract_doc(n)
+
     def parse(self):
         with open(self.filename, 'r') as f:
             program = f.read()
 
         self.ast = tree_from_tokens(list(tokenize(program)))
-        self.visit(self.ast)
+        # self.visit(self.ast)
+
+        for e in self.extract_doc(self.ast):
+            logger.debug(e)
